+++ ios/theolive/THEOplayerRCTView+THEOliveConfig.swift
@@ -6,7 +6,6 @@ import THEOplayerSDK
 struct THEOliveConfig {
     var externalSessionId: String?
     var discoveryUrl: String?
-    var useLegacyTHEOlive: Bool = false
 }
 
 extension THEOplayerRCTView {
@@ -15,7 +14,6 @@ extension THEOplayerRCTView {
         if let theoLiveConfig = configDict["theoLive"] as? NSDictionary {
             self.theoliveConfig.externalSessionId = theoLiveConfig["externalSessionId"] as? String
             self.theoliveConfig.discoveryUrl = theoLiveConfig["discoveryUrl"] as? String
-            self.theoliveConfig.useLegacyTHEOlive = theoLiveConfig["useLegacyTHEOlive"] as? Bool ?? false
         }
     }
 }