+++ src/api/source/SourceDescription.ts
@@ -62,18 +62,6 @@ export enum SourceIntegrationId {
 }
 
 /**
- * The playback pipeline to use to play a stream.
- *
- * @category Source
- * @platform android
- * @public
- */
-export enum PlaybackPipeline {
-  LEGACY = 'legacy',
-  MEDIA3 = 'media3'
-}
-
-/**
  * Describes the configuration of a player's source.
  *
  * @category Source
@@ -364,16 +352,6 @@ export interface BaseSource {
    * <br/> - Ignored for non-HLS streams.
    */
   hls?: HlsPlaybackConfiguration;
-
-  /**
-   * The playback pipeline to use for this stream.
-   *
-   * @defaultValue [PlaybackPipeline.MEDIA3]
-   * @since React Native THEOplayer SDK v9.0.0.
-   * @since Native THEOplayer SDK v9.0.0.
-   * @platform android
-   */
-  playbackPipeline?: PlaybackPipeline;
 }
 
 /**