+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [10.0.0] - 25-09-15
+
+### Added
+
+- Added support for player SDK v10. More info on the [migration documentation](./doc/migrating-to-react-native-theoplayer-10.md) page.
+- Added iOS and Android support for THEOads API and THEOads events.
+
+### Fixed
+
+- Fixed an issue on Android where the app could still enter picture-in-picture presentation mode after closing the player.
+
+### Changed
+
+- Upgraded the example app to React-Native v0.81.
+- All child views of `<THEOplayerView>` are being hidden when transitioning to picture-in-picture presentation mode on Android.
+
 ## [9.10.0] - 25-08-19
 
 ### Fixed